Make it compile in recent Linux systems.
This change makes a bunch of updates, and also some tweaks to make it work
under Clang (not the default).

The md5 generation code is failing tests now -- need to investigate a new
implementation I think.  Note that this code WILL BREAK other backup sets
because of this problem!

+++ qt/backup2/label_history_dlg.cpp
@@ -20,7 +20,7 @@ LabelHistoryDlg::LabelHistoryDlg(
 
   for (BackupItem item : history_) {
     QStringList view_strings;
-    QString size = QLocale().toString(item.size);
+    QString size = QLocale().toString((qulonglong)item.size);
     view_strings.append(item.date.toString());
     view_strings.append(item.type);
     view_strings.append(size);
