+++ qt/backup2/verify_helper.cpp
@@ -216,13 +216,13 @@ void VerifyHelper::SwitchToPage3() {
   vector<uint64_t> needed_volumes = restore_model_->GetNeededVolumes();
 
   ui_->verify_info_num_files->setText(
-      QLocale().toString(file_list.size()));
+      QLocale().toString((qulonglong)file_list.size()));
   ui_->verify_info_uncompressed_size->setText(
-      QLocale().toString(size));
+      QLocale().toString((qulonglong)size));
 
   QStringList volume_list;
   for (uint64_t volume : needed_volumes) {
-    volume_list.append(QLocale().toString(volume));
+    volume_list.append(QLocale().toString((qulonglong)volume));
   }
   ui_->verify_info_needed_volumes->setText(
       QLocale().createSeparatedList(volume_list));
@@ -264,11 +264,11 @@ void VerifyHelper::OnHistoryLoaded() {
           0, Qt::DisplayRole).toString());
   ui_->verify_backup_info_type->setText(item.type);
   ui_->verify_backup_info_size_uncompressed->setText(
-      QLocale().toString(item.size));
+      QLocale().toString((qulonglong)item.size));
   ui_->verify_backup_info_unique_size->setText(
-      QLocale().toString(item.unique_size));
+      QLocale().toString((qulonglong)item.unique_size));
   ui_->verify_backup_info_size_compressed->setText(QLocale().toString(
-      item.compressed_size));
+      (qulonglong)item.compressed_size));
   ui_->verify_date_description->setText(
       item.date.toString() + ": " + item.description + " (" + item.type + ")");
